working on StudentController

+++ database/migrations/2020_03_02_083052_create_addresses_table.php
@@ -16,8 +16,8 @@ class CreateAddressesTable extends Migration
         Schema::create('addresses', function (Blueprint $table) {
             $table->bigIncrements('id');
             //$table->timestamps();
-            //$table->increments('id');
-            $table->unsignedInteger('student_id');
+            //$table->unsignedInteger('student_id');
+            $table->unsignedBigInteger('student_id');
             $table->string('street_name', 250)->nullable();
             $table->string('street_number',250)->nullable();
             $table->integer('zip',false,true)->nullable();
